+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.0
+
+* Skip checks by creating files at `.medic/skipped`
+
+## 0.2.0
+
+* Load local checks from `.medic/checks`
+
 ## 0.1.0
 
 * Initial release